multiple changes working towards issue #9
* using eclipse-aether
* using JSR-330
* upgrading dependencies
* requiring maven 3.1.0 or greater
* adding new RepositoryDownloader for downloading from a maven repo

+++ src/main/java/com/github/klieber/phantomjs/download/WebDownloader.java
@@ -34,33 +34,37 @@ public class WebDownloader implements Downloader {
 
   private static final Logger LOGGER = LoggerFactory.getLogger(WebDownloader.class);
 
-
   private static final String DOWNLOADING = "Downloading phantomjs binary from {}";
   private static final String UNABLE_TO_DOWNLOAD = "Unable to download phantomjs binary from ";
 
   private final String baseUrl;
+  private final File target;
 
-  public WebDownloader(String baseUrl) {
+  public WebDownloader(String baseUrl, File target) {
     this.baseUrl = baseUrl;
+    this.target = target;
   }
 
   @Override
-  public void download(PhantomJSArchive archive, File target) throws DownloadException {
-    String url = buildDownloadUrl(archive);
-    try {
-      URL downloadLocation = new URL(url);
+  public File download(PhantomJSArchive archive) throws DownloadException {
+    if (!this.target.exists()) {
+      String url = buildDownloadUrl(archive);
+      try {
+        URL downloadLocation = new URL(url);
 
-      LOGGER.info(DOWNLOADING, url);
-      FileUtils.copyURLToFile(downloadLocation, target);
+        LOGGER.info(DOWNLOADING, url);
+        FileUtils.copyURLToFile(downloadLocation, target);
 
-      if (target.length() <= 0) {
-        throw new DownloadException(UNABLE_TO_DOWNLOAD+url);
+        if (target.length() <= 0) {
+          throw new DownloadException(UNABLE_TO_DOWNLOAD+url);
+        }
+      } catch (MalformedURLException e) {
+        throw new DownloadException(UNABLE_TO_DOWNLOAD+url, e);
+      } catch (IOException e) {
+        throw new DownloadException(UNABLE_TO_DOWNLOAD+url, e);
       }
-    } catch (MalformedURLException e) {
-      throw new DownloadException(UNABLE_TO_DOWNLOAD+url, e);
-    } catch (IOException e) {
-      throw new DownloadException(UNABLE_TO_DOWNLOAD+url, e);
     }
+    return this.target;
   }
 
   private String buildDownloadUrl(PhantomJSArchive archive) {
